# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Twitter-like API specification with Firebase Auth
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UserProfile {
  uid: string;
  username: string;
  displayName: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  avatarUrl: string | null;
  followersCount: number;
  followingCount: number;
  postsCount: number;
  createdAt: string;
}

export interface CreateUserProfileBody {
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Post {
  id: string;
  authorId: string;
  authorUsername: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl: string | null;
  content: string;
  likesCount: number;
  commentsCount: number;
  createdAt: string;
}

export interface CreatePostBody {
  content: string;
}

export interface PaginatedPosts {
  posts: Post[];
  /** @nullable */
  nextCursor: string | null;
  hasMore: boolean;
}

export interface FollowResult {
  following: boolean;
  targetUserId: string;
}

export interface FollowEntry {
  uid: string;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface FollowList {
  users: FollowEntry[];
}

export type GetUserPostsParams = {
  limit?: number;
  /**
   * @nullable
   */
  cursor?: string | null;
};

export type GetFeedParams = {
  limit?: number;
  /**
   * @nullable
   */
  cursor?: string | null;
};