# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Twitter-like API specification with Firebase Auth
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user profile
 */
export const GetMeResponse = zod.object({
  uid: zod.string(),
  username: zod.string(),
  displayName: zod.string(),
  bio: zod.string().nullable(),
  avatarUrl: zod.string().nullable(),
  followersCount: zod.number(),
  followingCount: zod.number(),
  postsCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Create user profile after signup
 */
export const CreateUserProfileBody = zod.object({
  username: zod.string(),
  displayName: zod.string(),
  bio: zod.string().nullish(),
  avatarUrl: zod.string().nullish(),
});

/**
 * @summary Get user profile by ID
 */
export const GetUserProfileParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserProfileResponse = zod.object({
  uid: zod.string(),
  username: zod.string(),
  displayName: zod.string(),
  bio: zod.string().nullable(),
  avatarUrl: zod.string().nullable(),
  followersCount: zod.number(),
  followingCount: zod.number(),
  postsCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Fetch all posts by a specific user
 */
export const GetUserPostsParams = zod.object({
  userId: zod.coerce.string(),
});

export const getUserPostsQueryLimitDefault = 20;

export const GetUserPostsQueryParams = zod.object({
  limit: zod.coerce.number().default(getUserPostsQueryLimitDefault),
  cursor: zod.coerce.string().nullish(),
});

export const GetUserPostsResponse = zod.object({
  posts: zod.array(
    zod.object({
      id: zod.string(),
      authorId: zod.string(),
      authorUsername: zod.string(),
      authorDisplayName: zod.string(),
      authorAvatarUrl: zod.string().nullable(),
      content: zod.string(),
      likesCount: zod.number(),
      commentsCount: zod.number(),
      createdAt: zod.coerce.date(),
    }),
  ),
  nextCursor: zod.string().nullable(),
  hasMore: zod.boolean(),
});

/**
 * @summary Get list of followers for a user
 */
export const GetUserFollowersParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserFollowersResponse = zod.object({
  users: zod.array(
    zod.object({
      uid: zod.string(),
      username: zod.string(),
      displayName: zod.string(),
      avatarUrl: zod.string().nullable(),
    }),
  ),
});

/**
 * @summary Get list of users a user is following
 */
export const GetUserFollowingParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserFollowingResponse = zod.object({
  users: zod.array(
    zod.object({
      uid: zod.string(),
      username: zod.string(),
      displayName: zod.string(),
      avatarUrl: zod.string().nullable(),
    }),
  ),
});

/**
 * @summary Create a new text post
 */
export const CreatePostBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Get a single post by ID
 */
export const GetPostParams = zod.object({
  postId: zod.coerce.string(),
});

export const GetPostResponse = zod.object({
  id: zod.string(),
  authorId: zod.string(),
  authorUsername: zod.string(),
  authorDisplayName: zod.string(),
  authorAvatarUrl: zod.string().nullable(),
  content: zod.string(),
  likesCount: zod.number(),
  commentsCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a post
 */
export const DeletePostParams = zod.object({
  postId: zod.coerce.string(),
});

/**
 * @summary Follow a user
 */
export const FollowUserParams = zod.object({
  targetUserId: zod.coerce.string(),
});

export const FollowUserResponse = zod.object({
  following: zod.boolean(),
  targetUserId: zod.string(),
});

/**
 * @summary Unfollow a user
 */
export const UnfollowUserParams = zod.object({
  targetUserId: zod.coerce.string(),
});

export const UnfollowUserResponse = zod.object({
  following: zod.boolean(),
  targetUserId: zod.string(),
});

/**
 * @summary Get personalized feed based on who the user follows
 */
export const getFeedQueryLimitDefault = 20;

export const GetFeedQueryParams = zod.object({
  limit: zod.coerce.number().default(getFeedQueryLimitDefault),
  cursor: zod.coerce.string().nullish(),
});

export const GetFeedResponse = zod.object({
  posts: zod.array(
    zod.object({
      id: zod.string(),
      authorId: zod.string(),
      authorUsername: zod.string(),
      authorDisplayName: zod.string(),
      authorAvatarUrl: zod.string().nullable(),
      content: zod.string(),
      likesCount: zod.number(),
      commentsCount: zod.number(),
      createdAt: zod.coerce.date(),
    }),
  ),
  nextCursor: zod.string().nullable(),
  hasMore: zod.boolean(),
});