# lib/api-zod/src/generated/types/followList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Twitter-like API specification with Firebase Auth
 * OpenAPI spec version: 0.1.0
 */
import type { FollowEntry } from "./followEntry";

export interface FollowList {
  users: FollowEntry[];
}