# lib/api-zod/src/generated/types/followResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Twitter-like API specification with Firebase Auth
 * OpenAPI spec version: 0.1.0
 */

export interface FollowResult {
  following: boolean;
  targetUserId: string;
}