# lib/api-zod/src/generated/types/paginatedPosts.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Twitter-like API specification with Firebase Auth
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from "./post";

export interface PaginatedPosts {
  posts: Post[];
  /** @nullable */
  nextCursor: string | null;
  hasMore: boolean;
}